lavf/rkrga: format map support
With this option, user can set dynamically the the output format of
rkrga_scale & rkrga_vpp filters, according to the input format. User
gives a pixel format map separated with + and and coupled with - sign.
Ie: formatmap=nv15-nv12+nv20-nv16 means that if the in input pixel
format is nv15, the output must be nv21, or if the input format is nv20
then the output format must be nv16. The mapping can be extended as
wished.

fmt1=fmt2:fmt3=fmt4... s specifically because players already serialize
the filter inputs with the same/similar syntax, therefore parsing of
individual avoptions together with dict type avoption is not always
possible due this restriction.